v2.5 : feat(audit): Add security audit and risk assessment document
refactor(controller): Update endpoint paths for ATS analysis, resume rewriting, and chat

fix(controller): Ensure student retrieval does not return null

refactor(model): Change ID type from Long to Integer for Job and User models

fix(service): Update FASTAPI base URL for AI service

fix(client): Remove unnecessary parameter from parseAndFilter method call

chore(scripts): Convert backend start script from bash to batch for Windows compatibility

+++ backend/src/main/java/com/careeros/model/Job.java
@@ -9,7 +9,7 @@ public class Job {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String title;
@@ -59,8 +59,8 @@ public class Job {
         this.fetchedAt = LocalDateTime.now();
     }
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getTitle() { return title; }
     public void setTitle(String title) { this.title = title; }
